+++ src/main/java/Command/AddCommand.java
@@ -1,3 +1,14 @@
+package duke.command;
+
+import duke.core.Ui;
+import duke.core.Storage;
+import duke.helper.DukeException;
+import duke.helper.DateTimeHelper;
+import duke.task.TaskList;
+import duke.task.ToDo;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
 import java.time.LocalDateTime;
 
 public class AddCommand extends Command {
