+++ src/main/java/Command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.core.Ui;
+import duke.core.Storage;
+import duke.helper.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
+
 public class DeleteCommand extends Command {
 
     public DeleteCommand(String inputCommand){
