+++ src/main/java/Core/Ui.java
@@ -1,3 +1,6 @@
+package duke.core;
+
+import duke.task.Task;
 import java.util.Scanner;
 import java.util.ArrayList;
