+++ src/main/java/Task/Event.java
@@ -1,3 +1,6 @@
+package duke.task;
+
+import duke.helper.DateTimeHelper;
 import java.time.LocalDateTime;
 
 public class Event extends Task{
